reformed the code and packaged the code
