+++ requirements.txt
@@ -1,0 +1,2 @@
+matplotlib==3.5.3
+numpy==1.21.6
