+++ setup.cfg
@@ -1,0 +1,13 @@
+[metadata]
+name="mananij_diffusion2D"
+version="0.0.3"
+author="Jayesh Manani"
+url="https://github.com/Simulation-Software-Engineering/diffusion2d"
+description="This code solves the diffusion equation using the Finite Difference Method."
+
+[options]
+packages = find:
+install_requires =
+  numpy
+  matplotlib
+  python_version>"3.6"
