+++ src/mananij_diffusion2D.egg-info/requires.txt
@@ -1,0 +1,2 @@
+numpy
+matplotlib
